% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day.R
\name{dayUI}
\alias{dayUI}
\title{嘉好按天合并对账单}
\usage{
dayUI(
  tabTitle = "按天合并",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = dayUI_left,
  func_right = dayUI_right,
  func_bottom = dayUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
嘉好按天合并对账单
}
\examples{
dayUI()
}
