% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{organizationUI}
\alias{organizationUI}
\title{嘉好对账单-日期组织}
\usage{
organizationUI(
  tabTitle = "日期组织",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = organizationUI_left,
  func_right = organizationUI_right,
  func_bottom = organizationUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
嘉好对账单-日期组织
}
\examples{
organizationUI()
}
