% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research.R
\name{researchUI}
\alias{researchUI}
\title{嘉好对账单-日期组织}
\usage{
researchUI(
  tabTitle = "研发订单",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = researchUI_left,
  func_right = researchUI_right,
  func_bottom = researchUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
嘉好对账单-日期组织
}
\examples{
researchUI()
}
